+++ packages/react/src/components/multi-factor-auth/multi-factor-auth.test.tsx
@@ -7,11 +7,11 @@ import {
 } from "../../context/test-slash-id-provider";
 import { createTestUser, inputEmail, inputPhone } from "../test-utils";
 
-describe.skip("#MultiFactorAuth", () => {
+describe("#MultiFactorAuth", () => {
   test("MFA flow", async () => {
-    const testUser = createTestUser()
-    const logInMock = vi.fn(() => Promise.resolve(testUser));
-    const mfaMock = vi.fn(() => Promise.resolve(testUser));
+    const testUser = createTestUser({ authMethods: ['email_link'] })
+    const logInMock = vi.fn(async () => testUser);
+    const mfaMock = vi.fn(async () => testUser);
     const user = userEvent.setup();
     const mfaText = "TEST TITLE MFA";