+++ packages/react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @slashid/react
 
+## 1.10.0
+
+### Minor Changes
+
+- 5e8d505: Move middleware exports to root
+- 4629fa8: Added the "label" option to OIDC factor config to enable overriding the UI label
+
+## 1.9.0
+
+### Minor Changes
+
+- 9e4ae09: Move the theme props to the SlashIDProvider
+- 433a8e4: Add the DynamicFlow component
+- 1d0d64c: Add <OrganizationSwitcher>, useOrganizations()
+- 1d0d64c: Add middleware concept, defaultOrganization middleware
+
 ## 1.8.3
 
 ### Patch Changes